feat: organize cypress tests
